+++ spec/helper/PlayerEventEmitter.ts
@@ -149,6 +149,13 @@ export class PlayerEventEmitter {
     });
   }
 
+  fireSourceLoadedEvent(): void {
+    this.fireEvent<PlayerEventBase>({
+      timestamp: Date.now(),
+      type: PlayerEvent.SourceLoaded,
+    });
+  }
+
   fireSourceUnloadedEvent(): void {
     this.fireEvent<PlayerEventBase>({
       timestamp: Date.now(),
@@ -300,7 +307,7 @@ export class PlayerEventEmitter {
         id,
         label,
       },
-    }as AudioTrackEvent);
+    } as AudioTrackEvent);
   }
 
   fireAudioRemovedEvent(id: string): void {